requirements.txt
asgiref==3.5.2
boto3==1.26.27
botocore==1.29.27
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
ratelim==0.1.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
